+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### __Breaking Changes__
 
 ### Added
+- Vue-directive: Added a vue-directive "bs-tooltip.js" to handle Bootstrap-Tooltips.
 
 ### Changed
 
@@ -20,6 +21,7 @@
 - Migrator: Fixed creation of index.html.
 - Gazetteer: Restored correct display of geographicalIdentifier as search result, if available.
 - SearchBar: Fixed unstable highlighting of 3D tiles at address coordinates.
+- LayerPills: The tooltip is always removed when a LayerPill is closed and no longer stays open on mouse click.
 
 ---